iot-producer/requirements.txt
#/Users/matteo/Documents/Lambda_IoT/iot-producer/requirements.txt

cassandra-driver
hdfs
requests
websocket-client
kafka-python